feat: initialize backend project modules, controllers, and core infrastructure with NestJS

+++ src/modules/categories/categories.service.ts
@@ -1,12 +1,21 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { CategoryRepository } from './repositories/category.repository';
 import { CreateCategoryDto } from './dto/request/create-category.dto';
 import { UpdateCategoryDto } from './dto/request/update-category.dto';
 import { CategoryEntity } from './entities/category.entity';
+import { ProductDetailRepository } from '../products/repositories/product-detail.repository';
+import { ProductRepository } from '../products/repositories/product.repository';
 
 @Injectable()
 export class CategoriesService {
-  constructor(private readonly categoryRepository: CategoryRepository) {}
+  constructor(
+    private readonly categoryRepository: CategoryRepository,
+    private readonly productRepository: ProductRepository,
+  ) {}
 
   async findAll(): Promise<CategoryEntity[]> {
     return this.categoryRepository.find();
@@ -27,7 +36,10 @@ export class CategoriesService {
     return this.categoryRepository.save(category);
   }
 
-  async update(id: number, updateCategoryDto: UpdateCategoryDto): Promise<CategoryEntity> {
+  async update(
+    id: number,
+    updateCategoryDto: UpdateCategoryDto,
+  ): Promise<CategoryEntity> {
     const category = await this.findById(id);
     this.categoryRepository.merge(category, updateCategoryDto);
     return this.categoryRepository.save(category);
@@ -35,6 +47,14 @@ export class CategoriesService {
 
   async remove(id: number): Promise<void> {
     const category = await this.findById(id);
+    // kiểm tra xem category còn sản phẩm nào không nếu còn thì không cho xóa
+    const product = await this.productRepository.find({
+      where: { categoryId: id },
+    });
+    
+    if (product.length > 0) {
+      throw new BadRequestException('Không thể xóa danh mục khi còn sản phẩm');
+    }
     await this.categoryRepository.remove(category);
   }
 }
